requirements-dev.txt
# Development & Testing Dependencies

# Testing Framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Test Data Generation
factory-boy==3.3.0

# E2E Testing
playwright==1.40.0
pytest-playwright==0.4.4

# Code Quality (optional, for future)
# flake8==6.1.0
# black==23.12.0
# isort==5.13.2
